refactor: extract generic func to get item by semver
